Function sunrise_ahci::pci::discover

fn discover() -> Vec<PciDevice>

Discover all pci devices, by probing the PID-VID of every slot on every bus.

A device is discovered when its (bus, slot, function 0x00)[register 0x00] != 0xFFFF_FFFF. Then, an additional PciDevice will be returned for every of its other functions that also return anything different from 0xFFFF_FFFF.